pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fusion-control-environment"
version = "0.1.0"
description = "HFM competition environment for fusion control"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
    "gymnasium>=0.29",
    "pyyaml>=5.0",
]

[project.optional-dependencies]
baseline = ["stable-baselines3>=2.0", "pyyaml>=5.0", "onnx>=1.16", "torch>=2.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["environment*"]
